Support multiple player properties files.
This allows separate maintenance of different parts of the configuration.
For example, settings that can be publicly visible can be maintained
in one location, while credentials and such are maintained elsewhere.

Files with names 'player.properties' and 'player.private.properties'
are loaded. The plain 'players.properties' file is loaded first, allowing
'player.private.properties' to add to override the 'player.properties'
values. At least one of the two files must be present.