Support multiple player properties files.
This allows separate maintenance of different parts of the configuration.
For example, settings that can be publicly visible can be maintained
in one location, while credentials and such are maintained elsewhere.

Files with names ending in 'player.properties' are loaded. The
plain 'players.properties' file is loaded first; other files are
then loaded in lexical order of their file name.